+++ src/aui-alert/js/aui-alert.js
@@ -40,6 +40,12 @@ A.Alert = A.Base.create('alert', A.Widget, [
      */
     renderUI: function() {
         this._uiSetCloseable(this.get('closeable'));
+
+        if (this.get('useARIA')) {
+            this.plug(A.Plugin.Aria);
+
+            this.aria.setAttribute('hidden', false, this.get('boundingBox'));
+        }
     },
 
     /**
@@ -76,6 +82,10 @@ A.Alert = A.Base.create('alert', A.Widget, [
     _onClickBoundingBox: function(event) {
         if (event.target.test('.' + CSS_CLOSE)) {
             this.hide();
+
+            if (this.get('useARIA') && !this.get('destroyOnHide')) {
+                this.aria.setAttribute('hidden', true, this.get('boundingBox'));
+            }
         }
     },
 
@@ -170,6 +180,19 @@ A.Alert = A.Base.create('alert', A.Widget, [
         destroyOnHide: {
             validator: A.Lang.isBoolean,
             value: false
+        },
+
+        /**
+        * Boolean indicating if use of the WAI-ARIA Roles and States should be enabled..
+        *
+        * @attribute useARIA
+        * @default true
+        * @type {Boolean}
+        */
+        useARIA: {
+            validator: A.Lang.isBoolean,
+            value: true,
+            writeOnce: 'initOnly'
         }
     },