+++ src/aui-sortable/js/aui-sortable-list.js
@@ -83,6 +83,8 @@ var SortableList = A.Component.create({
                 var dropNode = drop.get('node');
                 var dropOn = instance.get('dropOn');
 
+                dropNode.attr('dropzone', 'move');
+
                 return dropNode.one(dropOn);
             },
             validator: isFunction
@@ -199,6 +201,12 @@ var SortableList = A.Component.create({
 
             if (nodes) {
                 instance.addAll(nodes);
+
+                nodes.each(
+                    function(item) {
+                        item.attr('draggable', true);
+                    }
+                );
             }
         },